Simplify protocol query object
